Move city backfill into the rake task, not the Address model (#3123)
Per review: a one-off backfill shouldn't live permanently on the model. Drop Address#backfill_city! and the needs_city_backfill scope; inline the postcode lookup directly in the addresses:backfill_city task, matching the sibling addresses:regeocode task. Replaces the model specs with a rake-task spec (postcodes.io stubbed via WebMock).
